services/ingestion-fastapi/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-json-logger==2.0.7
prometheus-client==0.21.0
redis==5.0.8
aiokafka==0.10.0
PyJWT==2.9.0
httpx==0.27.2
pytest==8.3.2
pytest-asyncio==0.24.0
